+++ plugins/voxelman/dbg/plugininfo.d
@@ -1,0 +1,6 @@
+module voxelman.dbg.plugininfo;
+enum id = "voxelman.dbg";
+enum semver = "0.6.0";
+enum deps = [];
+enum clientdeps = [];
+enum serverdeps = [];